Issue #208 - Plugin restructure and cleanup
