+++ ait/core/server/plugins/__init__.py
@@ -1,0 +1,3 @@
+from data_archive import *
+from limit_monitor import *
+from openmct import *
