+++ config/leapseconds.dat
@@ -4,7 +4,7 @@ p1
 cdatetime
 datetime
 p2
-(S'\x07\xe3\x06\x1c\x00\x00\x00\x00\x00\x00'
+(S'\x07\xe4\x06\x1c\x00\x00\x00\x00\x00\x00'
 p3
 tp4
 Rp5
