+++ doc/source/ait.core.server.plugins.data_archive.rst
@@ -1,0 +1,7 @@
+ait.core.server.plugins.data\_archive module
+============================================
+
+.. automodule:: ait.core.server.plugins.data_archive
+    :members:
+    :undoc-members:
+    :show-inheritance:
