+++ doc/source/ait.core.server.plugins.limit_monitor.rst
@@ -1,0 +1,7 @@
+ait.core.server.plugins.limit\_monitor module
+=============================================
+
+.. automodule:: ait.core.server.plugins.limit_monitor
+    :members:
+    :undoc-members:
+    :show-inheritance:
