+++ doc/source/ait.core.server.plugins.rst
@@ -6,6 +6,8 @@ Submodules
 
 .. toctree::
 
+   ait.core.server.plugins.data_archive
+   ait.core.server.plugins.limit_monitor
    ait.core.server.plugins.openmct
 
 Module contents
